server/requirements.txt
fastapi>=0.99.1
uvicorn==0.22.0
SQLAlchemy==2.0.9
psycopg2-binary==2.9.6
python-dotenv==1.0.0
passlib==1.7.4
pyjwt==2.6.0
openai>=1.0.0